refactor: change naming to adequate to name convetion
