+++ src/config/interfaces/app-config.ts
