+++ src/config/interfaces/auth-config.ts
