+++ src/config/interfaces/index.ts
@@ -1,7 +1,7 @@
-export * from './appConfig';
-export * from './authConfig';
-export * from './databaseConfig';
-export * from './databaseConfig';
+export * from './app-config';
+export * from './auth-config';
+export * from './database-config';
+export * from './database-config';
 export * from './default';
-export * from './introductionConfig';
-export * from './thingDescriptionEventsConfig';
+export * from './introduction-config';
+export * from './thing-description-events-config';
