+++ src/config/interfaces/thing-description-events-config.ts
